compat: split stress into heavy (log/kv) and light (tpch1) targets
Heavy log/kv tests peak ~150 GB each under msan; with the previous single
stress target (cpu:32) the local scheduler ran up to 3 heavy chunks in
parallel (~450 GB demand on a 283 GB runner) -> OOM-kill of ydb nodes
(exit -9), e.g. test_log co-running with a 142 GB neighbor at 274/283 GB.

Move test_log/test_kv into a dedicated stress_heavy target with cpu:96 so
at most one 150 GB chunk runs at a time (memory -> cpu throttle, since
REQUIREMENTS(ram) is not honored locally). Keep the light tpch1 workload in
stress with cpu:48 so it still parallelizes. Update RECURSE and move the
log/kv mutes to the stress_heavy path across all sanitizer configs.
