+++ .github/config/muted_ya.txt
@@ -52,9 +52,9 @@ ydb/tests/compatibility/indexes test_bloom_filter_index.py.TestBloomFilterIndex.
 ydb/tests/compatibility/indexes test_fulltext_index.py.TestFulltextIndex.test_fulltext_index[rolling_stable-26-1-1_to_stable-26-2-1]
 ydb/tests/compatibility/indexes test_fulltext_index.py.TestFulltextIndex.test_fulltext_index[rolling_stable-26-2-1_to_prestable-26-3]
 ydb/tests/compatibility/workloads test_simple_reader.py.TestSimpleReaderTabletTransfer.test_simple_reader_tablet_transfer[rolling_stable-26-2-1_to_prestable-26-3]
-ydb/tests/compatibility/stress test_stress.py.TestStress.test_log[mixed_prestable-26-3_and_stable-26-2-1-column]
-ydb/tests/compatibility/stress test_stress.py.TestStress.test_log[mixed_stable-26-2-1-row]
-ydb/tests/compatibility/stress test_stress.py.TestStress.test_log[mixed_stable-26-2-1_and_stable-26-1-1-row]
+ydb/tests/compatibility/stress_heavy test_stress.py.TestStress.test_log[mixed_prestable-26-3_and_stable-26-2-1-column]
+ydb/tests/compatibility/stress_heavy test_stress.py.TestStress.test_log[mixed_stable-26-2-1-row]
+ydb/tests/compatibility/stress_heavy test_stress.py.TestStress.test_log[mixed_stable-26-2-1_and_stable-26-1-1-row]
 ydb/tests/compatibility/stress test_stress.py.TestStress.test_tpch1[mixed_prestable-26-3_and_stable-26-2-1-column-date64]
 ydb/tests/compatibility/stress test_stress.py.TestStress.test_tpch1[mixed_prestable-26-3_and_stable-26-2-1-column]
 ydb/tests/compatibility/workloads test_table_schema_compatibility.py.TestTableSchemaCompatibilityRestart.test_create_alter_table_on_one_version_read_on_another[restart_prestable-26-3_to_prestable-26-3]
