+++ ydb/tests/compatibility/ya.make
@@ -4,6 +4,7 @@ RECURSE(
     indexes
     topics
     stress
+    stress_heavy
     rolling
     udf
     federated_queries
